Warnings as errors, builder for CallbackInfra
The main build file now has tighter compiler options for reporting all
warnings and treating warnings as errors.

The acceptance test doesn't depent directly on the
CallbackInfrastructure class but instead depends on the
CallbackInfrastrctureBuilder to create an instance of it. This way the
concern of building the system and using the system are separated.