Dedupe CMakeLists.txt. Separate implementation
Refactor the various CMakeLists.txt files into one master file at the
root of the library which in-turn builds the acceptance, integration and
unit tests.

Separate the implementation from the interface header file. Put
implementation files in the src subdirectory. Build solution as a lib
and link with (test) executables.